daemon: extract fabric/VRF/interface reconcile from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the fabric
IPVLAN reconcile, VRF reconcile, management-VRF interface rebind, and
interface reconcile move verbatim into daemon_apply_interfaces.go.

Moved declarations (byte-identical bodies, no rename, no logic change):
applyFabricIPVLAN, applyVRFReconcile, rebindManagementVRFIfaces,
applyInterfaceReconcile. Same package, no caller changes; per-file
imports pruned by goimports. Build: go build ./pkg/daemon/... green.
